refactor(react): split state.ts into state.get/state.use modules
State.get and State.use become self-assigning modules (state.get.ts /
state.use.ts) that wire themselves onto State; host.ts loads them for their
side effect. The standalone use(subject) hook moves to host.ts alongside the
other Runtime hooks, and host.ts absorbs the agnostic barrel (State, context,
use) that state.ts used to provide. The /state subpath repoints to host.js
(unchanged for preact and external consumers); state.get.js/state.use.js are
marked sideEffects so the self-assignment survives bundling.

Tests realign 1:1: state.get.test.tsx (State.get), state.use.test.tsx
(State.use), and the standalone-use tests move into host.test.ts.
