+++ packages/react/package.json
@@ -23,7 +23,9 @@
   "type": "module",
   "sideEffects": [
     "./dist/index.js",
-    "./dist/state.js",
+    "./dist/host.js",
+    "./dist/state.get.js",
+    "./dist/state.use.js",
     "./src/jsx-runtime.ts"
   ],
   "main": "dist/index.js",
@@ -34,8 +36,8 @@
       "default": "./dist/index.js"
     },
     "./state": {
-      "types": "./dist/state.d.ts",
-      "default": "./dist/state.js"
+      "types": "./dist/host.d.ts",
+      "default": "./dist/host.js"
     }
   },
   "engines": {
