+++ packages/react/src/host.ts
@@ -1,5 +1,10 @@
-import { Component, watch, unbind, observer } from '@expressive/mvc';
+import { Component, event, watch, unbind, observer } from '@expressive/mvc';
 import { provide, type Context } from './context';
+import './state.get';
+import './state.use';
+
+export { State } from '@expressive/mvc';
+export { Consumer, Provider, Context, provide } from './context';
 
 declare module '@expressive/mvc' {
   interface Component {
@@ -82,6 +87,51 @@ export function useHook<T = void>(
   }, []);
 
   return current.output;
+}
+
+/** Subscribe to an existing observable instance within a component. */
+export function use<T extends object>(subject: T) {
+  const { current } = Runtime.useRef<{
+    proxy: T;
+    source?: T;
+    mounted: number;
+    unwatch?: () => void;
+  }>({ mounted: 0, proxy: subject });
+
+  const update = Runtime.useState(() => current.mounted++)[1];
+
+  if (current.source !== subject) {
+    const status = observer(subject);
+
+    if (status === undefined)
+      throw new Error('Provided object is not observable.');
+
+    current.unwatch?.();
+    current.source = subject;
+
+    if (status === null) {
+      current.unwatch = undefined;
+      current.proxy = subject;
+    } else {
+      if (!status.ready) event(subject);
+
+      let init = true;
+
+      current.unwatch = watch(subject, (next, changed) => {
+        current.proxy = next;
+        if (changed.length && !init)
+          update((x) => x + 1);
+      });
+
+      init = false;
+    }
+  }
+
+  Runtime.useEffect(() => () => {
+    if (--current.mounted < 1) current.unwatch?.();
+  }, []);
+
+  return current.proxy;
 }
 
 function bootstrap(this: Component, context: Context){
